Added
Added custom exceptions for easier handling of messages and codes.
Do not know if it is the right way to use the codes, but they could work well in an API.

Began tinkering on a db-class, but realise that it might be the wrong way to do this.
Better to check for the indivdual storages instead. This might be removed.